Sinusoidal shake revised
All screen shake is now vector based, much more reliable and accurate.

Made everything based on frame ticks in order to support actual frequency calculations.

Completely rewrote and simplified the fader.

This will work with everything, but I continued to do the sin() first because it is the most complex.